% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_ds.R
\name{accuracy_ds}
\alias{accuracy_ds}
\title{accuracy_ds}
\usage{
accuracy_ds(x)
}
\arguments{
\item{x}{data frame with class atribute and the data that has been taken
into regionalization}
}
\value{
the accuracy measure of random forest classification
}
\description{
Calculates the accuracy measure based on caret's ranger model for the regionalization done by polygon_ds or points_ds
}
