% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5ab_ds_accuracy.R
\name{ds_accuracy}
\alias{ds_accuracy}
\title{ds_accuracy}
\usage{
ds_accuracy(data.to.accu)
}
\arguments{
\item{data.to.accu}{data frame with class attribute and the data that has been taken
into regionalization}
}
\value{
the accuracy measure of random forest classification
}
\description{
Calculates the accuracy measure based on caret's ranger model for the regionalization done
by ds_polygon or ds_points. It represents how accurate would random forest model predict
appropriate community based only on their attributes.
}
