% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b_ds_polygon.R
\name{ds_polygon}
\alias{ds_polygon}
\title{ds_polygon

Creates a vector of community assignment based on neighboring polygons. It
creates a topological structure in which nodes represent polygons and the edge
is the similarity between nodes. Communities are created using fast greedy
algorithm that maximizes their modularity.}
\usage{
ds_polygon(
  x,
  k = 2,
  queen = TRUE,
  data = -grep(names(x), pattern = "^geom"),
  similarity.measure = "euclidean",
  style = "B",
  disjoint = FALSE,
  n.neigh = 8,
  plot = TRUE,
  explain = TRUE
)
}
\arguments{
\item{x}{point or polygon shapefile data;}

\item{k}{number of clusters;}

\item{queen}{if TRUE, a single shared boundary point meets the contiguity condition,
if FALSE, more than one shared point is required; note that more than one shared boundary
point does not necessarily mean a shared boundary line}

\item{data}{attributes of the spatial data frame to calculate similarity or
distance measure;}

\item{similarity.measure}{Character or function to declare distance
method transformed into similarity measure. If method is
character, method must be "mahalanobis" or "euclidean", "maximum",
"manhattan", "canberra", "binary" or "minkowski". If method is one of
"euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski",
see dist for details, because this function as used to compute the distance.
If method="mahalanobis", the mahalanobis distance is computed between
neighbor areas. If method is a function, this function is used to compute
the distance.}

\item{style}{style can take values “W”, “B”, “C”, “U”, “minmax” and “S”}

\item{disjoint}{if default settings generate error occurring to disjoint
subgraphs it means, that in some places points or polygons are to disjoint
to generate one connected graph. Use disjoint = T to enforce that one graph
will be created. This is a slower option.}

\item{n.neigh}{number of neighbors considered in the k-nearest neighbor
algorithm that builds topology}

\item{plot}{should the neighborhood be plotted}

\item{explain}{logical. If TRUE a machine learning (randomForest
using 5 fold cross validation) model is being constructed based
on the data provided for regionalization. The accuracy of this model
explains how much of the regionalization can be attributed to the data
and how much to the spatial distribution.}
}
\value{
vector of numbers representing regions to which each element belongs to
}
\description{
ds_polygon

Creates a vector of community assignment based on neighboring polygons. It
creates a topological structure in which nodes represent polygons and the edge
is the similarity between nodes. Communities are created using fast greedy
algorithm that maximizes their modularity.
}
