% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1c_find_no_clusters.R
\name{find_no_clusters}
\alias{find_no_clusters}
\title{find_no_clusters}
\usage{
find_no_clusters(
  x,
  polygon = TRUE,
  queen = TRUE,
  method = "euclidean",
  data = 2:ncol(x),
  style = "B",
  n.neigh = 8,
  disjoint = FALSE,
  range = 2:30
)
}
\arguments{
\item{x}{spatial data for regionalization}

\item{polygon}{logical if the data is polygon or point}

\item{queen}{if data is polygon and without disjoint polygons, should the
neighbourhood be treated by queen topology or rook topology}

\item{method}{Character or function to declare distance method. If method
is character, method must be "mahalanobis" or "euclidean", "maximum",
"manhattan", "canberra", "binary" or "minkowisk". If method is one of
"euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowisk",
see dist for details, because this function as used to compute the
distance. If method="mahalanobis", the mahalanobis distance is computed
between neighbour areas. If method is a function, this function is used to
compute the distance.}

\item{data}{data to analyze similarity between regions}

\item{style}{style can take values “W”, “B”, “C”, “U”, “minmax” and “S”}

\item{n.neigh}{number of neighbours considered in the k-nearest neighbour
algorithm that builds topology}

\item{disjoint}{if default settings generate error occuring to disjoint
subgraphs it means, that in some places points or polygons are to disjoint
to generate one connected graph. Use disjoint = T to enforce that one graph
will be created. This is a slower option.}

\item{range}{number of divisions to test the modularity. THe biger the
numbers, the longer it will take to calculate plot}
}
\value{
A vector of modularity measures for given range of divisions
}
\description{
Helps identify how many regions should be divided by analyzing the changes in modularity value
}
