% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4ab_part_communities.R
\name{part_communities}
\alias{part_communities}
\title{part_communities}
\usage{
part_communities(k, fg)
}
\arguments{
\item{k}{number of clusters to create regionalization}

\item{fg}{hierarchical community object created from \code{build_graph()}}
}
\value{
a vector of classes - numbers of regions that particular
polygon or point are classified to
}
\description{
Divides graph into defined number of regions
}
