% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1a_points_ds.R
\name{points_ds}
\alias{points_ds}
\title{points_ds}
\usage{
points_ds(
  x,
  k = 2,
  queen = TRUE,
  data = -grep(names(x), pattern = "^geom"),
  method = "euclidean",
  style = "B",
  disjoint = FALSE,
  n.neigh = 8,
  plot = TRUE,
  accuracy = TRUE
)
}
\arguments{
\item{x}{point or polygon shapefile data;}

\item{k}{number of communities;}

\item{queen}{if TRUE, a single shared boundary point meets the contiguity condition,
if FALSE, more than one shared point is required; note that more than one shared boundary
point does not necessarily mean a shared boundary line}

\item{data}{atributes of the spatial data frame to calculate similarity or
distance measure;}

\item{method}{Character or function to declare distance method. If method is
character, method must be "mahalanobis" or "euclidean", "maximum",
"manhattan", "canberra", "binary" or "minkowisk". If method is one of
"euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowisk",
see dist for details, because this function is used to compute the distance.
If method is set to "mahalanobis", the mahalanobis distance is computed between
neighbour points. If method is a function, this function is used to compute
the distance.}

\item{style}{style can take values “W”, “B”, “C”, “U”, “minmax” and “S” (see spdep::nb2listw)}

\item{disjoint}{if default settings generate error occuring to disjoint
subgraphs it means, that in some places points or polygons are to disjoint to
generate one connected graph. Use disjoint = T to enforce that one graph will
be created. This is a slower option.}

\item{n.neigh}{number of neighbours considered in the k-nearest neighbour
algorithm that builds topology network}

\item{plot}{should the neighbourhood be plotted}

\item{accuracy}{logical should accuracy be calculated based on randomForest algorithm}
}
\value{
vector of numbers representing regions to whicheach element
}
\description{
Creates a vector of community assignment based on neighbouring points. It is based
on a topological network structure where points represent nodes and the edges
are the degree of similarity between those nodes. Communities are created using fast greedy
algorithm that maximizes their modularity.
}
\examples{
data("realEstate")
realEstate <- points_ds(realEstate,k = 5, )
\dontrun{
library(dismo)
pol <- voronoi(xy = as(realEstate,'Spatial'))
qtm(pol, "class")
}

}
