% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2b_prepare_polygons.R
\name{prepare_polygons}
\alias{prepare_polygons}
\title{prepare_polygon}
\usage{
prepare_polygons(x, queen, disjoint, n.neigh, plot)
}
\arguments{
\item{x}{point object}

\item{queen}{logical should the queen or the rook neighborhood be calculated}

\item{disjoint}{logical if polygons are not continuous}

\item{n.neigh}{number of nearest neighbors that should be taken into
consideration for building graph if disjoint==TRUE}

\item{plot}{logical if TRUE a plot showing neighborhoods is being presented}
}
\value{
neighborhoods for community finding
}
\description{
Prepares polygons for regionalization - calculates
neighborhood objects
}
