% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{realEstate}
\alias{realEstate}
\title{Real Estate dataset}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 250 rows and 6 columns.
}
\usage{
data(realEstate)
}
\description{
Location of real estate transactions in Poznań between 2007-2013 with some
landscape characteristics. All of the attributes represent the euclidean distance to some
insttutions or amenities.
}
\examples{
data(realEstate)
}
\keyword{datasets}
