% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{socioGrid}
\alias{socioGrid}
\title{Socio-economic dataset}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 250 rows and 6 columns.
}
\usage{
data(socioGrid)
}
\description{
Socio-economic factors aggregated to a grid of 500m squares
}
\examples{
data(socioGrid)
}
\keyword{datasets}
