% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3ab_build_graph.R
\name{build_graph}
\alias{build_graph}
\title{build_graph}
\usage{
build_graph(x, data, x.nb, method, style)
}
\arguments{
\item{x}{polygon or point data from witch neighborhood object will be created}

\item{data}{data frame to build weights between polygons/points}

\item{x.nb}{neighborhood object created by prepare_points() or prepare_polygons()}

\item{method}{method to calculate similarity/distance between neighboring points or polygons}

\item{style}{style of neighborhood (see `spdep::nb2listw``)}
}
\value{
a list containing community object ("fg") and graph object ("graph")
}
\description{
Creates fastgreedy community graph for extracting regions
}
