% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dspace-package.R
\docType{package}
\name{dspace-package}
\alias{dspace}
\alias{dspace-package}
\title{dspace: Dividing Space - Data Driven Segementation of Spatial Data}
\description{
Divides geographical polygon and point data into spatially cohesive 
    regions based on fast greedy community finding algorithm.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dabrowskia/dspace}
  \item Report bugs at \url{https://github.com/dabrowskia/dspace/issues}
}

}
\author{
\strong{Maintainer}: Adam Dabrowski \email{adam.dabrowski@amu.edu.pl}

Authors:
\itemize{
  \item Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})
}

}
\keyword{internal}
