% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2c_plot_modularity.R
\name{plot_modularity}
\alias{plot_modularity}
\title{plot_modularity}
\usage{
plot_modularity(mod)
}
\arguments{
\item{mod}{modularity object}
}
\value{
a plot of modularity values calculated for a given range of clusters
}
\description{
plots the modularity statistic against the number of clusters.
}
\examples{
data("quakes")
data("socioGrid")
modularity <- find_no_clusters(socioGrid, disjoint = TRUE, n.neigh = 6)
plot_modularity(modularity)
}
