% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_modularity.R
\name{plot_modularity}
\alias{plot_modularity}
\title{plot_modularity}
\usage{
plot_modularity(mod)
}
\arguments{
\item{mod}{modularity object}
}
\value{
a plot of modularity values calculated for a given range of clusters
}
\description{
plots the modularity statistic against the number of clusters.
}
