% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a_prepare_points.R
\name{prepare_points}
\alias{prepare_points}
\title{prepare_points}
\usage{
prepare_points(x, method, n.neigh, plot)
}
\arguments{
\item{x}{point object of class sf}

\item{method}{the distance/similarity to calculate}

\item{n.neigh}{at least how many neighbors should be taken into
consideration}

\item{plot}{logical if TRUE a plot showing neighborhoods
is being presented}
}
\value{
neighborhoods for community finding
}
\description{
Prepares points for regionalization - changes simple features to
SpatialPolygonsDataFrame and calculates neighborhood objects
}
