% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2b_prepare_polygons.R
\name{prepare_polygons}
\alias{prepare_polygons}
\title{prepare_polygon}
\usage{
prepare_polygons(x, queen, method, disjoint, n.neigh, plot)
}
\arguments{
\item{x}{point object}

\item{queen}{logical should the wueen or the rook neighbourhood be calculated}

\item{method}{the distance/similarity to calculate}

\item{disjoint}{logical if polygons are not continuous}

\item{n.neigh}{at least how many neighbours should be taken into regionalization if disjoint==T}

\item{plot}{logical if TRUE a plot showing neighbourhoods is beeing presented}
}
\value{
neighbourhoods for coummunity finding
}
\description{
Prepares polygons for regionalization - changes simplea features to SpatailPolygonsDataFrame an calculates neighbourhood objects
}
